Move create_session from test__internal to util
